+++ src/forms/index.ts
@@ -1,5 +1,3 @@
-// FormField
-
 // ArrayField
 export { ArrayField, type ArrayFieldProps } from "./array-field";
 // CheckboxField
@@ -27,7 +25,14 @@ export { DirtyCounter, type DirtyCounterProps } from "./dirty-counter";
 export { DirtyFormGuard, type DirtyFormGuardProps } from "./dirty-form-guard";
 // FileField
 export { FileField, type FileFieldProps } from "./file-field";
+// FormField
 export { FormField, type FormFieldProps } from "./form-field";
+// FormMarkers
+export {
+	type FormMarkers,
+	FormMarkersProvider,
+	type FormMarkersProviderProps,
+} from "./form-markers";
 // InlineEdit
 export {
 	InlineEdit,